Add an example doc folder for writing and testing a Mill plugin (#3416)
Most of the interesting stuff is in
`example/extending/plugins/7-writing-mill-plugins/build.sc`. It's a
simple example build for a Mill plugin that has unit, integration, and
example tests

I did some unrelated refactoring as part of this PR, to try and reduce
the messiness of the example folder and make it better match the current
docsite, and some cleanup in the `build.sc` that was necessary to make
the example test pass (e.g. `dev.*` is now `dist.*`)