Use shorter partially-qualified names when labelling named tasks during planning (#3418)
Rather than `foo.super.bar.qux.Baz.foo`, we now use `foo.super.Baz`. If
there are multiple `Baz` classes with the same task being overriden
(probably rare), we use the longest partially-qualified path suffix
that's able to uniquely distinguish them. We also always drop the
right-most path segment, since all that does is repeat the task name we
already have on the left

Also tweaked the `visualize` tasks to use the same planning/labelling
logic as the evaluator, so the labels we see in the SVG are the same
labels seen in the terminal

Adjusted existing tests, and added additional tests to cover the
partially-qualified-name selection