Set `forkEnv` to an empty Map
Since os-lib propagates the system environment by default when it runs
subprocesses, setting `forkEnv` to the system environment, is redundant,
but also confusing, since removing a key from the Map doesn't stop mill
to pass the environment variable to the run program.
This sets a new default, which is a semantic breaking change, so it should
go in Mill 0.12.0, which is a binary compatible but behavior incompatible
release.