Be stricter about requiring `build.` prefix for cross-`package.mill`-file references (#3456)
This ensures that IDEs can trace the execution since we pretend that
everything lives in `package build`. Previously, if the `build.` prefix
is elided, we get code that works but the IDE does the wrong thing. Now
that code is a compile error and you need to use `build.` which the IDE
can understand