Favor -SNAPSHOT version when publishing Mill locally
This makes the millVersion task return a more stable value. That value
ends up in BuildInfo files, so having it not change prevents it
trigerring re-compilations.

Without this change, adding files to the staging area, committing
things, or sometimes just creating new files, was changing the
millVersion, and as a consequence trigerring a re-compilation of many
things.