Break docs `pages/` folder into subfolders and add forwarder aliases (#3629)
The module hierarchy now matches the `example/` folder hierarchy, which
should make things easier to navigate. Shortened the individual `.adoc`
file names to use a shorter convention to remove duplication with the
folder name

It seems that the `:page-aliases:` declaration needs to happen on the
line immediately after the title, and no later. Otherwise it silently
doesn't work, which was causing a lot of forwarder pages to be missing
before. Now they seem to appear properly