Use build info for versions in kotlinlib (#3633)
Some minor cleanup for the previous work on the Kotlin-related modules:

* Read tooling versions from build info
(https://github.com/com-lihaoyi/mill/pull/3620#discussion_r1779445633
and
https://github.com/com-lihaoyi/mill/pull/3620#discussion_r1779445811)
* Declare versions only once in the unit tests
* Change parent module to `JavaModule` for Ktfmt support, because we
don't need the code to be compiled in this module, so declaring Kotlin
version there was looking a bit strange. In fact only `sources()` method
is needed, which is available in the `JavaModule`. In the future
`KotlinModule` may be split into different traits: the one which
collects Kotlin+Java files, but doesn't do any compilation (so Kotlin
version is not needed) and the one which actually does compilation.

---------

Co-authored-by: 0xnm <0xnm@users.noreply.github.com>