Add some BSP integration tests (#3608)
This adds some BSP integration tests, ensuring the Mill BSP server
answers correctly basic BSP requests. BSP responses are serialized on
disk and committed, and we compare the requests we get to those on disk.
These fixtures on disk can be automatically generated and updated by
changing a variable (currently, `BspServerUtil.updateFixtures`), so that
it's quite convenient to generate them, and update them and get a nice
diff of the changes at the same time.