Fix JavaModule.artifactTypes (#3719)
Follow-up of https://github.com/com-lihaoyi/mill/pull/3703 that added
`JavaModule.artifactTypes`. `JavaModule` performs two dependency
resolutions: one for `compileClasspath`, the other for `runClasspath`.
https://github.com/com-lihaoyi/mill/pull/3703 used `artifactTypes` in
the `runClasspath` one, but not in the `compileClasspath` one. The PR
here fixes that, and adds unit tests for both paths.

This addresses the issues found around
https://github.com/com-lihaoyi/mill/pull/3696#issuecomment-2405891608.