Add a default config for logback
Since we bundle the logback dependency, it runs with a built-in default config printing all messages with at least DEBUG level on the console. This can be very verbose.

This change includes a default config creating a log file `out/mill.log` in the current workind directory. Since Mill server  now always runs in a sandbox dir, this should not polute the users working directory.

If a user is want to change the log configuration, she ca simple provide an alternatve config with `-Dlogback.configurationFile=/path/to/config.xml`