Allow `.mill-jvm-opts` to interpolate environment variables, add `.mill-opts` (#3841)
This PR extends `.mill-jvm-opts` to support interpolating JVM variables.
This provides a bit of flexibility in allowing `$PWD` and `$USER` to be
used when passing `-Dproperty=` to Mill. This could in theory be fleshed
out even further to run arbitrary shell snippets, but for now I this
should provide enough flexibility to meet the immediate needs.

Also added a `.mill-opts`, which is similar to `.mill-jvm-opts` but used
for Mill arguments rather than JVM arguments. This provides a way to
persistent required Mill configuration in a format that is similar to
what we already have today (CLI args). This would be a common place to
put things like `--jobs=0.5C` or `--allow-positional` or
`--no-build-lock`.

Updated the integration tests and documentation

related discussions

* https://github.com/com-lihaoyi/mill/discussions/1226 
* https://github.com/com-lihaoyi/mill/discussions/3840