Swap out page-version switcher for global version switcher (#3852)
Previously we disabled the global one because it was redundant, but it
seems the page version switcher doesn't play well with renamed pages
even when page aliases are defined, resulting in a lot of page version
links being broken. So for now we just disable the page version switcher
in favor of the global one

Tested locally using `docs.githubPages`. The link to 0.11.12 is still
broken, but it should be fixed once we release 0.11.13 with
https://github.com/com-lihaoyi/mill/commit/c61d1eda0a574703530b7d1abd37f12f9c1e46ab
included