Share profile loggers between nested evaluations (#3885)
We broaden the scope of `profileLogger`/`chromeProfileLogger` from
within `def evaluate` to instead last as long as the `Evaluator` itself,
and share it with any sub-`Evaluator`s created e.g. to redirect output
during `show`

`Evaluator` is now `AutoCloseable`, and needs to be closed after being
used in `MillBuildBootstrap` and in `UnitTester`

For now I decided not to share the profilers between different
evaluation levels of the meta-build, since they already don't overlap:
one goes to `out/`, the parent to `out/mill-build/`, etc. We could do
that if we think it would be useful to give people a consolidated
profile, but it might be a bit more work (e.g. we would need to properly
prefix entries in the profile to distinguish the different meta levels).

Added a unit test to ensure the `show compile` mill-profile and
mill-chrome-profile contain the expected keys for both `show` and for
`compile` and some of its upstream tasks

Fixes https://github.com/com-lihaoyi/mill/issues/3879