Add nice error when non-exclusive task depends on exclusive task (#3887)
Fixes https://github.com/com-lihaoyi/mill/issues/3875

I placed a check to raise an error when a non-exclusive task depends on
an exclusive task.

We could place this check any time from task instantiation to just
before the `NoSuchElementException` gets thrown. I ended up putting it
as late as possible, such that the failure is localized, which allows
other un-related parts of the build to proceed despite the failure, as
well as allowing other errors in the build to be reported in parallel
(rather than having to re-run the build over and over after fixing each
one to see the next one)

Covered with an integration test that asserts on the error message. Best
reviewed with Ignore Whitespace