JMH Plugin : This is a forkArg not a compiler option (#3957)
#3578 set out to allow passing JVM flags (for example
`--add-modules=jdk.incubator.vector`) into the JVM used by the JMH
plugin when trying to benchmark something in mill.

That implementation kind of works, - it uses `javaCoptions` to pass to
the JVM args though, which isn't very idiomatic mill. Runtime arguments
are usually passed as `forkArgs`. This PR seeks to use 'forkArgs'
instead of the 'javaCoptions' of the mill module.

Both are a list of strings in the end, so both work, but I think this
revised version more correctly follows mill nomenclature as forkedArgs
rather than compiler flags.

Apologies for the spamminess.