Allow users to customize coursier ResolutionParams (#3949)
This adds a `JavaModule#resolutionParams` task, allowing users to
customize coursier's
[`ResolutionParams`](https://github.com/coursier/coursier/blob/b04937dda13cee80b3743c84f471a69eb430047e/modules/coursier/shared/src/main/scala/coursier/params/ResolutionParams.scala).
These allow to customize some aspects of dependency resolution.

In particular, `ResolutionParams#enableDependencyOverrides`, when set to
`Some(false)`, disables the recent BOM changes added (and enabled by
default) in coursier `2.1.17`. This might be useful, if ever users
aren't happy with the versions it selects, or want to wait a bit before
enabling that, or if they find out performance suffers because of it for
them.

The new task is anonymous (it's not a target, it's not serialized on
disk), as `ResolutionParams` has a field that doesn't easily lend itself
to serialization (`rules`, because of the `Rule` type, which is an open
abstract class).