Some improvements to pythonlib (#3992)
This is in view of #3928, specifically some essential things required
for the first task, `example/pythonlib/basic/`.

You can look at commits individually if you like. The gist of this PR is
to improve the scaffolding we currently have to unblock the previously
mentioned tasks. It can be summarized by the following points (roughly
each corresponding to one commit):

- [x] make `run` task the default, following the same convention used in
`ScalaModule` and `JavaModule`
- [x] add a command to run an interactive REPL. This is named `console`,
again following the conventions of scalalib
- [x] rework the way source files are handled:
- support multiple source directories, again following similar
conventions of scalalib
- instead of aggreting python scripts in one syntetic directory, keep
them where they are and instead manipulate PYTHONPATH or pass in
parameters to various directories where appropriate
- [x] define a module for unit tests, similar to the TestModule of
scalalib