Integrate lefou/`millw` and `millw.bat` into Mill repository, fix GenIdea on windows (#4052)
Fixes https://github.com/com-lihaoyi/mill/issues/2161, fixes
https://github.com/com-lihaoyi/mill/issues/3203, parts of
https://github.com/com-lihaoyi/mill/issues/4050

For now I just merged the two files unchanged except for tweaking the
comment to refer to the `Original Project page`. We can continue to
evolve the scripts post-merge. We already had a copy in `ci/mill.bat`
that I removed, and changed the github actions to point at the top-level
version, and integrated the new `mill.bat` into the doc-site download
instructions and example zips

The changes to the `mill`/`mill.bat` scripts should be covered by the
existing integration and bootstrap tests



This is enough to make Mill's own codebase importable to IntelliJ on
Windows